chore: move transtornos para modulo de paciente
