--- apps/api/src/main/java/br/org/apae/api/disorder/domain/repository/DisorderRepository.java
@@ -1,9 +1,0 @@
-package br.org.apae.api.disorder.domain.repository;
-
-import br.org.apae.api.disorder.domain.model.Disorder;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface DisorderRepository extends JpaRepository<Disorder, UUID>{
-}
